separating componetns in react
